(*)Cancel Z_to_H factors in MOM_diabatic_aux.F90
  Cancelled out factors of GV%Z_to_H in MOM_diabatic_aux.F90 to simplify the
code and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.
This involved changing the units of 6 internal variables in
diagnoseMLDbyDensityDifference to use thickness units or related units.  In one
place, GV%Rho0 was replaced with GV%H_to_RZ.  Because GV%Z_to_H is an exact
power of 2 in Boussinesq mode, all answers are bitwise identical in that mode,
but in non-Boussinesq mode this conversion involves multiplication and division
by GV%Rho_0, so while all answers are mathematically equivalent, this change
does change diagnostics at roundoff in non-Boussinesq mode.