*Cancel out Z_to_H factors in MOM_bkgnd_mixing.F90
  Cancelled out factors of GV%Z_to_H in MOM_bkgnd_mixing.F90 to simplify the
code and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.
This involved changing the units of 3 internal variables in compute_ddiff_coeffs
and one element in the bkgnd_mixing_cs type to use thickness units or their
inverse.  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all
answers are bitwise identical in that mode, but in non-Boussinesq mode this
conversion involves multiplication and division by GV%Rho_0, so while all
answers are mathematically equivalent, this change does change answers at
roundoff in non-Boussinesq mode.