*Cancel Z_to_H in MOM_lateral_mixing_coeffs.F90
  Cancelled out factors of GV%Z_to_H in MOM_lateral_mixing_coeffs.F90 to
simplify the code and reduce the dependence on the value of GV%Rho_0 in
non-Boussinesq mode.  This involved changing the units of one internal variable
in calc_QG_Leith_viscosity to use inverse thickness units (as its descriptive
comment already indicated). The variable N2 is no longer calculated separately
in calc_slope_functions_using_just_e, but this code is left in a comment as it
may be instructive.  The previously incorrect units of one comment were also
fixed. Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers
are bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.