*Cancel out Z_to_H factors in MOM_tidal_mixing.F90
  Cancelled out factors of GV%Z_to_H in MOM_tidal_mixing.F90 to simplify the
code and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.
This involved changing the units of 2 internal variables in
calculate_CVMix_tidal, 9 internal variables in add_int_tide_diffusivity, 2
elements of the tidal_mixing_diags type and 3 elements of the tidal_mixing_cs
type to use thickness units or related units. Because GV%Z_to_H is an exact
power of 2 in Boussinesq mode, all answers are bitwise identical in that mode,
but in non-Boussinesq mode this conversion involves multiplication and division
by GV%Rho_0, so while all answers are mathematically equivalent, this change
does change answers at roundoff in non-Boussinesq mode.