+*Cancel Z_to_H factors in MOM_set_diffusivity.F90
  Cancelled out factors of GV%Z_to_H in MOM_set_diffusivity.F90 to simplify the code
and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.  This
involved changing the units of numerous variables to use thickness units or related units,
including:

 - 4 internal variables in find_TKE_to_Kd
 - 4 internal variables in find_N2
 - 6 internal variables in add_drag_diffusivity
 - 7 internal variables in add_LOTW_BBL_diffusivity,
 - 5 internal variables in add_MLrad_diffusivity,
 - 4 internal variables in set_BBL_TKE
 - 1 internal variable in set_diffusivity_init
 - 1 element in the set_diffusivity_CS type

Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.