+Pass Hml to 10 routines in thickness units
  Changed the units of the mixed layer depth variables (Hml) passed to several
routines from height units to thickness units to simplify the code and reduce
the dependence on the value of GV%Rho_0 in non-Boussinesq mode.  The routines
that are impacted by this change are:
 - energetic_PBL_get_MLD
 - KPP_get_BLD
 - diabatic
 - diabatic_ALE
 - diabatic_ALE_legacy
 - layered_diabatic
 - bulkmixedlayer
 - calculate_CVMix_conv
 - call_tracer_column_fns
 - MOM_generic_tracer_column_physics

  Also changed the units of the Hml element of the MOM_control_struct to
thickness units, which can lead to a change of the description of the units of
the hML variable in the restart files if there is an active ice shelf, even in
Boussinesq mode.

  This change also required the addition of a verticalGrid_type argument to
energetic_PBL_get_MLD, and corresponding changes to the calls to this routine
from the diabatic driver, neutral diffusion and lateral boundary diffusion
codes.

  The internal units of the ML_depth element of energetic_PBL_CS and the MLD_io
argument to ePBL_column were also changed to thickness units, and GV%Rho0 was
replaced with the equivalent expression based on GV%H_to_RZ in one line of
ePBL_column.

  Also corrected the unit descriptions of two variables in comments in
MOM_bulk_mixed_layer.F90.

  Because this commit is just delaying the conversion from thickness units to
height units, all answers are bitwise identical in both Boussinesq and
non-Boussinesq cases.  However, there is a new argument to a publicly visible
routine, and the units of Hml arguments are changed in 10 public routines (but
not to any of the interfaces called from the driver level).