+wave_speed arg mono_N2_depth in thickness units
  Changed the units of the optional mono_N2_depth argument to wave_speed,
wave_speed_init and wave_speed_set_param in thickness units instead of height
units.  Accordingly, the units of one element each in the diagnostics_CS and
wave_speed_CS and a local variable in VarMix_init are also changed to thickness
units.  Because this is essentially just changing when the unit conversion
occurs, all answers are bitwise identical, but there are changes to the units of
an optional argument in 3 publicly visible routines.