*+Vertvisc_type boundary layers in thickness units
  Changed the units of the 5 boundary layer thickness elements of the
vertvisc_type ([tb]bl_thick_[uv] and MLD) from height units to thickness units
and cancelled out redundant factors of GV%H_to_Z to more accurately reflect the
nature of these fields, to simplify the code, and reduce the dependence on the
value of GV%Rho_0 in non-Boussinesq mode.  The MLD argument to
mixedlayer_restrat and the MLD_in argument to mixedlayer_restrat_general were
similarly rescaled to thickness units.  The units of restart variable MLD change
in non-Boussinesq mode as a result of these changes.

  New local variables incorporating the GV%H_to_Z scaling factor were added in
set_BBL_TKE and set_viscous_BBL for simplicity and efficiency, but this may
change the order of arithmetic for non-power-of-2 variables in non-Boussinesq
mode.  A few other local variables in these same routines that were renamed to
avoid misleading indications of their units. Additionally, the temporary
thicknesses and related variables used in the melt potential calculations in
extract_surface_state were also recast into thickness units.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.  This commit changes the units of 5 elements of a
widely-used transparent type and the units of 2 arguments in public subroutine
interfaces.