+*Work in thickness units in MOM_vertvisc.F90
  Revised the code in MOM_vertvisc.F90 to work internally with thickness-based
units for the viscous coupling between layers, which reduces the dependence on
the value of GV%Rho_0 in non-Boussinesq mode.  This involved changing the units
of numerous variables to use thickness units or related units, including:

 - 9 elements in the vertvisc_CS type
 - 1 argument to the private routine find_coupling_coef
 - 1 argument to the private routine find_coupling_coef_gl90
 - 1 argument ("a") to the public routine write_u_accel
 - 1 argument ("a") to the public routine write_v_accel
 - 1 internal variable in vertvisc
 - 1 internal variable in vertvisc_remnant
 - 8 internal variables in vertvisc_coef,
 - 7 internal variables in find_coupling_coef,

  Local variables that are no longer needed were eliminated in vertvisc and
vertvisc_remnant, while a new local variable was added to find_coupling_coef.
In 6 places the Boussinesq reference density was replaced with GV%H_to_RZ, which
is equivalent to the reference density in Boussinesq mode, but scales to 1 in
non-Boussinesq mode.

  The previous dimensional rescaling factor for KD_GL90 was incorrect (and
inconsistent with the correct scaling factor used when reading in the analagous
spatially varying kappa_gl90_2d); this has been corrected in this commit.

  To accomodate this change in vertvisc, the units of the "a" arguments to
write_u_accel and write_v_accel were also changed to use thickness-based
arguments.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.  This commit changes the units of 2 arguments in
public (diagnostic) subroutine interfaces.