*+Vertvisc_type Ray_[uv] in thickness flux units
  Changed the units of the 2 Rayleigh drag velocity elements (Ray_u and Ray_v)
of the vertvisc_type from vertical velocity units to thickness flux units and
cancelled out redundant factors of GV%H_to_Z to more accurately reflect the
nature of these fields, to simplify the code, and reduce the dependence on the
value of GV%Rho_0 in non-Boussinesq mode.  The units of the internal Rayleigh
variable in set_viscous_BBL were changed accordingly to thickness-based units.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.  This commit changes the units of 2 elements of a
widely-used transparent type and the units of 2 arguments in public subroutine
interfaces.