*Use thickness units in MOM_mixed_layer_restrat.F90
  Revised the code in MOM_mixed_layer_restrat.F90 to work internally with
thickness-based units for the restratification timescale calculation and other
internal calculations, which reduces the dependence on the value of GV%Rho_0 in
non-Boussinesq mode.  This involved changing the units of numerous variables to
use thickness units or related units, including:

 - The Kv_restrat element in the mixedlayer_restrat_CS type
 - 4 arguments and an internal variable in the private function growth_time
 - 5 internal variables in mixedlayer_restrat_general
 - 4 internal variables in mixedlayer_restrat_BML

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.