*Refactor MOM_thickness_diffuse.F90 into thickness
  Use thickness-based streamfunctions and other thickness-based internal
calculations in MOM_thickness_diffuse.  For example, the overturning
streamfunctions with this change are now in m3/s in Boussinesq mode, but kg/s in
non-Boussinesq mode.  These changes partially reduces the dependence on the
value of GV%Rho_0 in non-Boussinesq mode.  This change includes changing the
units of 11 internal variables in thickness_diffuse_full and 3 arguments to the
private routine streamfn_solver.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.