+*Work in thickness units in MOM_kappa_shear
  Refactored MOM_kappa_shear to differentiate between thicknesses and geometric
heights by working internally with dynamic viscosities and diffusivities in
non-Boussinesq mode, and to keep thicknesses in thickness units rather than
converting them to distances in height units.  Doing so reduces the influence of
the Boussinesq reference density in non-Boussinesq configurations, and it lays
the groundwork for a proper non-Boussinesq implementation of this module.

  This commit includes changing the units of 65 internal variables, some of
which were also renamed to reflect their new role, and adding 8 new internal
variables. There are 2 new arguments to find_kappa_tke (one is necessary to
differentiate thicknesses from vertical distances and while the other is used to
preserve Boussinesq-mode answers), and 1 new argument to kappa_shear_column.
There are 12 lines with obvious possibilities for simplification in
find_kappa_tke, but this would change answers; perhaps an ANSWER_DATE flag could
allow them at at some point.  The units of 14 internal subroutine arguments and
3 elements of the Kappa_shear_CS type change.  However the external interfaces
to this module are unaltered.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode unless GV%Rho_0 is chosen to be an integer power of 2.