+Use vertical distances in vertvisc_coef
  Use vertical distances when calculating the vertical viscous coupling
coefficients in vertvisc_coef and find_coupling_coef.  This includes a new dz
argument to vertvisc_coef and changes from [H ~> m or kg m-2] to [Z ~> m] for
the units of 4 arguments to find_coupling_coef, 1 argument to
find_coupling_coef_gl90, and 3 elements of the vertvisc_CS_type. The units of 29
internal variables are changed, and there are 5 new internal variables to
reflect that there are both thickness and height change variables needed in
vertvisc_coef.  A total of 27 conversion factors were cancelled out, while 9 new
ones were added. This change also required thickness_to_dz to be called in
several places in step_MOM_dyn_split_RK2, step_MOM_dyn_unsplit and
step_MOM_dyn_unsplit_RK2 to generate the vertical distances to pass to
vertvisc_coef. All answers are bitwise identical for Boussinesq test cases, but
there is a new argument to a public interface.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode unless GV%Rho_0 is chosen to be an integer power of 2.