(*)Use thickness_to_dz in MOM_kappa_shear
  Use thickness_to_dz to convert layer thicknesses to vertical distances in
Calculate_kappa_shear and Calc_kappa_shear_vertex.  The internal variable dz was
renamed to h_lay in these two routines for greater clarity, and there are
several other new internal variables to support this transition.  The results
are bitwise identical in Boussinesq mode, and mathematially equivalent in most
existing non-Boussinesq cases, but it will change answers in non-Boussinesq mode
if SEMI_BOUSSINESQ is set to false.