(*)Work with thickness and height in ePBL
  Work with both thickness (h) and height change (dz) variables in energetic_PBL
using thickness_to_dz to translate between the two.  In Boussinesq mode, these
are a simple rescaling by a constant factor, but in non-Boussinesq mode they
vary by a factor the density.  Some calculations are unnecessarily duplicated in
Boussinesq mode, so this might slow the model slightly, but they are not
duplicative in non-Boussinesq mode.  This change involves the addition of a new
argument to ePBL_column and changes to the units of two other arguments, changes
to the units of an element of the energetic_PBL_CS type, the addition of 5 new
internal variables, the removal of 1 internal variable, and changes to the units
of 19 internal variables (many of which were renamed to reflect their new
units). With this change, most of the remaining thickness to height conversion
factors have been eliminated, and Rho_0 is only used in Boussinesq mode.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change can change answers at roundoff in
non-Boussinesq mode unless GV%Rho_0 is chosen to be an integer power of 2.