+(*)Use thickness_to_dz in CVMix routines
  Use thickness_to_dz to convert thicknesses to vertical distances in
KPP_calculate, KPP_compute_BLD, calculate_CVMix_conv and calculate_CVMix_shear.
This requires the addition of a new thermo_var_ptr argument to KPP_calculate.
In addition, the internal calculations in MOM_CVMix_KPP were reverted to using
height units, as they are on the main branch of MOM6, with the net effect of
reducing the differences from the standard version.  All answers in Boussinesq
test cases are bitwise identical, but there is a new argument to KPP_calculate.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change can change answers at roundoff in
non-Boussinesq mode unless GV%Rho_0 is chosen to be an integer power of 2.