+(*)Change visc%bbl_thick_u back to depth units
  Returned to depth units for the top and bottom boundary layer thicknesses that
are stored in the vertvisc_type, and cancelled out scaling factors in several of
this impacted routines.  The units of 4 arrays in the transparent vertvisc_type
are changed, and there are new arguments to several subroutines.

  The most extensive refactoring occurs in the MEKE code, where there are
changes to the units of 2 elements in the MEKE_CS type. and arguments to
MEKE_equilibrium, MEKE_equilibrium_restoring, MEKE_lengthScales and
MEKE_lengthScales_0d.  The units of 8 internal variables in the MEKE code were
changed, and 3 variables eliminated.  A new verticalGrid_type argument was added
to MEKE_equilibrium_restoring and MEKE_init.

  In set_BBL_TKE, the units of two internal variables are changed, while there
is minor refactoring in set_viscous_ML, and the units of one internal variable
and one variable in the set_visc_CS type change.  However, in set_viscous_BBL
these changes are accommodated with newly added rescaling factors, and more
extensive revisions will be needed.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change can change answers at roundoff in
non-Boussinesq mode unless GV%Rho_0 is chosen to be an integer power of 2.