(*)Refactor set_viscous_BBL to be non-Boussinesq
  Refactored set_viscous_BBL and set_viscous_ML to work in height units rather
than thickness units in various places where it is more appropriate to do so
when in non-Boussinesq mode, and use a call to thickness_to_dz to convert
between thicknesses and vertical distances.  In some places this leads to extra
calculations using separate arrays that are rescaled duplicates of each other
in Boussinesq mode, which will probably slow the model down.  These changes are
extensive but localized to this file, with 3 new or altered elements of the
opaque set_visc_CS type, 18 new internal variables, and 35 existing internal
variables that have revised units or were renamed.  In particular one
extensively used thickness curvature variable that had been called just "a" was
renamed "crv" to more accurately reflect its purpose and to make it easier to
find when rescaling it, with similar changes to two other closely related
variables. These changes include the elimination of 17 rescaling factors
between thickness and height units were eliminated, as were two spots that
inappropriately (and unnecessarily, as it turns out) used GV%Rho0 even when in
non-Boussinesq mode.  The public interfaces are unchanged, and the answers in
Boussinesq mode are bitwise identical.  In non-Boussinesq mode, the answers are
unaltered when GV%Rho_0 is an exact power of 2, and are mathematically
equivalent otherwise.

  There is, however, one rescaling bug due to a hard-coded unrescaled
dimensional constant that was fixed.  It occurs when setting a tiny floor in a
thickness in one inverse calculation.  However, this appears to be so small
that it does not change any answers in the MOM6-examples test suite, and could
only do so if ANGSTROM is set to be many orders of magnitude less than the
typical value of 1e-10 m, so it was decided that no runtime bug-flag is needed
in this case.
1e-10 m, so it was decided that no runtime bug-flag is needed in this case.