+Revise units of visc%ustar_BBL to [H T-1]
  Changed the units of bottom friction velocity in visc%ustar_BBL from
[Z T-1 ~> m s-1] to [H T-1 ~> m s-1 or kg m-2 s-1] to avoid repeated unit
conversions, that effectively involve division and multiplication by the
Boussinesq Rho_0 when in non-Boussinesq mode. After this commit, there are a net
of 3 fewer conversion factors than before. In addition, a long-standing comment
questioning whether there is double-counting of tidal mixing has been addressed
(there is not) and the comment has been revised accordingly.  Answers are
bitwise identical in Boussinesq mode, and they are mathematically equivalent in
non-Boussinesq mode, but they can change if RHO_0 is not an integer power of 2.
There are changes to the units of an array in a transparent public type.