+(*)Refactor set_BBL_TKE to use depth units
  Refactored add_LOTW_BBL_diffusivity, add_MLrad_diffusivity, and set_BBL_TKE to
work in units of layer vertical extent rather than layer thickness to eliminate
16 instances of rescaling, and to give results in non-Boussinesq mode that have
reduced dependence on the Boussinesq reference density.  There is a new
thermo_vary_type argument to set_BBL_TKE.  The thickness argument to
add_MLrad_diffusivity (in [H ~> m or kg m-2] has been replaced with a vertical
extent argument (in [Z ~> m]).  There are 4 additional calls to thickness_to_dz
associated with these changes.  The units of 1 subroutine arguments and 14
internal variables are changed. There is 7 new internal variable, while 4
internal variables were eliminated. Answers are bitwise identical in Boussinesq
mode, and they are mathematically equivalent in non-Boussinesq mode, but they
can change if RHO_0 is not an integer power of 2.  There are changes to the
arguments to some routines.