+(*)Use thickness_to_dz in find_N2_bottom
  Use thickness_to_dz to convert thicknesses to vertical distances in
find_N2_bottom, and work in thickness units in that routine.  This leads to the
introduction of one new internal variable and changes to the units of two
others.  One unit thickness unit conversion factor was eliminated as a result
of this commit.  All answers are bitwise identical in Boussnesq mode, but
answers can change at roundoff in non-Boussinesq mode unless RHO_0 is an
integer power of 2.