+(*)Use depth units in calculate_tidal_mixing
  Use depth units in calculate_tidal_mixing, thereby reducing the dependency on
the Boussinesq reference density when the model in non-Boussinesq mode. This
includes replacing the 3-d layer thickness argument to calculate_tidal_mixing,
calculate_CVMix_tidal and add_int_tide_diffusivity with a 2-d slice of vertical
distances across layers.  There is also consolidated logic around the call to
thickness_to_dz in set_diffusivity.  This includes changing the unite of 3
variables in the opaque tidal_mixing_cs type, 2 arrays in the tidal_mixing_diags
type, and 9 internal variables (some of which were renamed to clarify their
purpose). All answers in Boussinesq test cases are bitwise identical, but there
is a change to the arguments to 3 routines, and answers can change at roundoff
in non-Boussinesq mode unless RHO_0 is an integer power of 2 even in
semi-Boussinesq mode.