+(*)Use depth units in full_convection
  Changed the units of the Kddt_smooth argument to full_convection and its
counterpart in smoothed_dRdT_dRdS to use units of [H Z ~> m2 or kg s-1], which
becomes a mass based diffusivity when in non-Boussinesq mode.  This change also
uses vertical distances for internal calculations in smoothed_dRdT_dRdS, and
includes a call to thickness_to_dz in the full_convection routine.  All answers
in Boussinesq test cases are bitwise identical, but there is a change to the
arguments to 2 routines, and answers can change at roundoff in non-Boussinesq
mode unless RHO_0 is an integer power of 2 even in semi-Boussinesq mode.