+visc%MLD and Hml arguments back in height units
  Changed the units of the MLD element of the vertvisc_type and the units of the
mixed layer depth variables (Hml) passed to 12 routines from thickness units
back to height units to simplify the code and reduce the scope of the changes
related to the non-Boussinesq mode conversions.

  The routines that are impacted by this change are:
 - energetic_PBL_get_MLD
 - KPP_get_BLD
 - diabatic
 - diabatic_ALE
 - diabatic_ALE_legacy
 - layered_diabatic
 - bulkmixedlayer
 - calculate_CVMix_conv
 - call_tracer_column_fns
 - MOM_generic_tracer_column_physics
 - mixedlayer_restrat
 - mixedlayer_restrat_general

  Also changed the units of the Hml element of the MOM_control_struct to
vertical distance units, which can lead to a reversion of the description of the
units of the hML variable in the restart files if there is an active ice shelf,
even in Boussinesq mode.  The units of restart variable MLD change in
non-Boussinesq mode as a result of these changes.

  This change also allowed for the removal of verticalGrid_type arguments to
energetic_PBL_get_MLD and KPP_get_BLD, and corresponding changes to the calls to
these routines from the diabatic driver, neutral diffusion and lateral boundary
diffusion codes.

  All answers in Boussinesq test cases are bitwise identical, but answers can
change at roundoff in non-Boussinesq mode unless RHO_0 is an integer power of 2.
This commit changes the units of 1 element of a widely-used transparent type and
the units of several arguments in public subroutine interfaces (but not to any
of the interfaces called from the driver level).