*Use forces%tau_mag in mixedlayer_restrat
  Use forces%tau_mag in place of forces%ustar in several places in the
mixedlayer_restrat module when in non-Boussinesq or semi-Boussinesq mode.  All
expressions are mathematically equivalent when in semi-Boussinesq mode and
answers do not change if Rho0 = 1024.0, but new expressions use tv%SpV_avg in
place of 1/Rho0 when in non-Boussinesq mode and answers will change in that
case.  The units of the ustar_min element of the opaque mixedlayer_restrat_CS
type are changed from [Z T-1] to [H T-1].  All answers in Boussinesq mode are
bitwise identical, but they change in some cases in non-Boussinesq mode.