*Use fluxes%tau_mag in add_MLrad_diffusivity
  Use fluxes%tau_mag in place of fluxes%ustar in add_MLrad_diffusivity when in
non-Boussinesq or semi-Boussinesq mode.  All expressions are mathematically
equivalent when in semi-Boussinesq mode and answers do not change if RHO_0 is
1024.0, but new expressions use tv%SpV_avg in place of 1/Rho0 when in
non-Boussinesq mode and answers will change in that case.  There is a new
thermo_var_ptrs type argument to the internal routine add_MLrad_diffusivity to
permit these changes.   All answers in Boussinesq mode are bitwise identical,
but they change in some cases in non-Boussinesq mode.