+*Use forces%tau_mag & tv%SpV_avg in vertvisc_coef
  When in non-Boussinesq mode, use forces%tau_mag and tv%SpV_avg instead of
forces%ustar and GV%Rho0 when interpolating the friction velocity and stress
magnitude in find_coupling_coef. This change includes the addition of a
thermo_vars_type argument to vertvisc_coef and find_coupling_coef.   All answers
in Boussinesq mode are bitwise identical, but they change in some cases in
non-Boussinesq mode, and there is a new mandatory argument to a publicly visible
subroutine interface.