*Use forces%tau_mag & tv%SpV_avg in bulkmixedlayer
  When in non-Boussinesq mode, use forces%tau_mag and tv%SpV_avg instead of
forces%ustar and GV%Rho0 to determine the surface TKE flux and the friction
velocity used in the bulk mixed layer code.  This has the effect of reducing the
dependence of the mixed layer code on the Boussinesq reference density when in
non-Boussinesq mode.  This change includes using a call to average_specific_vol
to translate the mass of the mixed layer into the mixed layer thickness, and it
includes the addition of a new vert_visc_type argument to find_starting_TKE.
This change includes the addition of a thermo_vars_type argument to
vertvisc_coef and find_coupling_coef.  All answers in Boussinesq mode are
bitwise identical, but they change in some cases in non-Boussinesq mode.