*Use forces%tau_mag & tv%SpV_avg in set_viscous_ML
  When in non-Boussinesq mode, use forces%tau_mag and tv%SpV_avg instead of
forces%ustar and GV%Rho0 to determine the friction velocity used in
set_visous_ML.  This includes a change to the units of one of the elements of
the opaque set_visc_CS type.  This has the effect of reducing the dependence of
the mixed layer code on the Boussinesq reference density when in non-Boussinesq
mode.  All answers in Boussinesq mode are bitwise identical, but they change in
some cases in non-Boussinesq mode.