*+Use fluxes%tau_mag & tv%SpV_avg in energetic_PBL
  When in non-Boussinesq mode, use forces%tau_mag and tv%SpV_avg instead of
forces%ustar and GV%Rho0 to determine the friction velocity and mechanical TKE
input used in energetic_PBL, and to convert the local TKE to a turbulent
velocity.  This includes the addition of a new energy input argument and an
interface specific volume argument to ePBL_column and the removal of an unused
argument to find_mstar.  This set of changes has the effect of reducing the
dependence of the mixed layer code on the Boussinesq reference density when in
non-Boussinesq mode.  All answers in Boussinesq mode are bitwise identical, but
they change in some cases in non-Boussinesq mode.