*Use specific volume in non-Boussinesq wave speeds
  Use specific volume derivatives to set the values of g_prime at interfaces
when in non-Boussinesq mode.  This eliminates the dependencies of the
non-Boussinesq wave speed calculations on the Boussinesq reference density, but
at the cost of about 180 lines of nearly duplicated code and the addition of 10
new internal variables. All answers in Boussinesq mode are bitwise identical,
but the estimated wave speeds change slightly in non-Boussinesq mode.