*+Add FULL_DEPTH_EADY_GROWTH_RATE
  Add the new runtime parameter FULL_DEPTH_EADY_GROWTH_RATE to indicate that the
denominator of an Eady growth rate calculation should be based on the full depth
of the water column, rather than the nominal depth of the bathymetry.  The new
option is only the default for fully non-Boussinesq cases. Also corrected the
case of some indices to follow the MOM6 soft convention using case to indicate
the staggering position of variables.  There is a new logical element in the
VarMix_CS type, and to accommodate one possible version of these changes, there
is a new thermo_var_type argument to calc_slope_functions_using_just_e.

  A primordial horizontal indexing bug was corrected in the v-direction slope
calculation.  Because it only applies for very shallow bathymetry, does not
appear to impact any existing test cases and went undetected for at least 12
years, it was corrected directly rather than wrapping in another new runtime
flag.  However, this bug is being retained for now in a comment to help with
review and debugging if the answers should change unexpectedly in some yet-to-be
identified configuration.

  All answers are bitwise identical in the existing MOM6-examples test suite, but
by default they will change when fully non-Boussinesq, and there is a new entry
in some MOM_parameter_doc files.