*Non-Boussinesq revision of KPP_compute_BLD
  Eliminated dependency on the Boussinesq reference density in KPP_compute_BLD
when in fully non-Boussinesq mode.  This includes the addition of the new array
with the buoyancy difference between the water in and below the mixed layer.
The answers will change in non-Boussinesq mode when the KPP boundary layer
parameterization is in use, but they are bitwise identical in all Boussinesq
test cases.