*Non-Boussinesq revision of calculate_CVMix_conv
  Remove dependency on the Boussinesq reference density in calculate_CVMix_conv
when in fully non-Boussinesq mode.  This includes the use of a layer thicknesses
rather than a vertical distance in the denominator of the calculation of the
buoyancy frequency and to the units of the internal rescaled gravity variable in
this routine when in non-Boussinesq mode.  The answers will change in
non-Boussinesq mode when the KPP convection parameterization is in use, but they
are bitwise identical in all Boussinesq test cases.