*Non-Boussinesq thickness_diffuse_full
  Refactored thickness_diffuse_full when in non-Boussinesq mode to avoid any
dependencies on the Boussinesq reference density, and to translate the volume
streamfunction into the mass streamfunction using an appropriately defined
in-situ density averaged to the interfaces at velocity points.  This form
follows the suggestions of Appendix A.3.2 of Griffies and Greatbatch (Ocean
Modelling, 2012) when in non-Boussinesq mode.

  These changes use a call to thickness_to_dz to set up a separate variable with
the vertical distance across layers, and in non-Boussinesq mode they use
tv%SpV_avg to estimate in situ densities.  The code changes are extensive with
14 new or renamed internal variables, and changes to the units of 4 other
internal variables.  After this change, GV%Rho, GV%Z_to_H and GV%H_to_Z are no
longer used in any non-Boussinesq calculations.  Because some calculations have
to be redone with the separate thickness and dz variables, this will be more
expensive than the original version.

  No public interfaces are changed, and all answers are bitwise identical in
Boussinesq or semiBoussinesq mode, but they will change in non-Boussinesq mode.