+Add find_rho_bottom
  Added the new subroutine find_rho_bottom to return a 1-d slice of the in situ
density averaged over a specified distance from the bottom when in fully non-
Boussinesq mode or an array filled with the Boussinesq reference density.  This
new routine is not yet used with this commit, but it has been tested in another
commit that will follow shortly.  All answers are bitwise identical, but there
is a new public interface.