+*Non-Boussinesq internal tide drag uses density
  Use the in situ near bottom density to calculate the internal tide drag,
energy input and energy loss terms when in non-Boussinesq mode.  This change
includes the addition of an argument containing the near-bottom density to
propagae_int_tide, itidal_lowmode_loss and find_N2_bottom. The recently added
routine find_rho_bottom is used to calculate this near-bottom density.  Several
instances where the Boussinesq reference density or GV%Z_to_H were used have
been eliminated from use in non-Boussinesq cases by this change. It also
includes the rescaling of a variable in int_tide_CS, and a new element with the
bottom drag in the int_tide_input_type. All answers are bitwise identical in
Boussnesq mode, but some solutions will change in non-Boussinesq mode with this
change, and there are new arguments to publicly visible subroutines and a new
element in a transparent type.