*Non-Boussinesq interface_filter
  Refactored interface_filter when in non-Boussinesq mode to avoid any
dependencies on the Boussinesq reference density, and to translate the volume
streamfunction into the mass streamfunction using an appropriately defined
in-situ density averaged to the interfaces at velocity points.  This form is
similar to the one that is used in thickness_diffuse.  This also adds a halo
update for tv%SpV_avg when filter_interface is about to be called because of
the use of relatively wide halos there.

  No public interfaces are changed, and all answers are bitwise identical in
Boussinesq or semiBoussinesq mode, but they will change in non-Boussinesq mode
cases that use the interface filter.