+Add tv%valid_SpV_halo to debug non-Boussinesq mode
  Added the integer valid_SpV_halo to the thermo_var_ptrs type to indicate
whether the SpV_array has been updated and its valid halo size, to facilitate
error detection and debugging in non-Boussinesq mode.  Tv%valid_SpV_halo is set
to the halo size in calc_derived_thermo or after a halo update is done to
tv%SpV_avg, and it is set to a negative value right after calls that change
temperatures and salinities (such as by ALE remapping) unless there is a call
to calc_derived_thermo.  Tests for the validity of tv%SpV_avg were added to
thickness_to_dz, find_rho_bottom and calc_isoneutral_slopes, with fatal errors
issued if invalid arrays would be used, but more tests could perhaps be used in
any parameterization routines where tv%SpV_avg is used directly.  Handling the
updates to tv%SpV_avg this way helps to avoid unnecessary calls to
calc_derived_thermo, which in turn has equation of state calls that can be
expensive, while also providing essential verification of new code related to
the non-Boussinesq code.  These tests can probably be commented out or removed
for efficiency once there is a full suite of regression tests for the fully
non-Boussinesq mode of MOM6.  One call to calc_derived_thermo was also added
just before the initialization call to ALE_regrid that will be needed with the
next commit, but does not change answers yet.  All answers are bitwise
identical, but there is a new element in a transparent and widely used type.