+*Non-Boussinesq form of mixedlayer_restrat_Bodner
  Modified mixedlayer_restrat_Bodner to avoid using the Boussinesq reference
density when in non-Boussinesq mode.  This includes changes to the units of
several variables, including wpup_filtered, which appears in
mixedlayer_restrat_CS and in the restart files generated with some mixed layer
restratification settings. There are 6 new internal variables, while units of 9
internal variables were changed, and several rescaling factors were cancelled
out or replaced.  When in non-Boussinesq mode, the mixed layer buoyancy
gradients are determine from the average specific volume referenced to the
surface pressure, rather than from the average potential density.

  The rescaling of several chksum calls for thicknesses was modified to
GV%H_to_mks to avoid any dependence on RHO_0.

  There are also comments where variable units were corrected or added.

  A logical branch was added based on the correct mask for land or OBC points to
avoid potentially ill-defined calculations of the magnitude of the Bodner
parameterization streamfunction, some which were leading to NaNs.

  Also set a tiny but nonzero default value for MIN_WSTAR2 to avoid NaNs in some
calculations of the streamfunction magnitude.

  Within mu, the expression for dd was revised in a mathematically equivalent
way to avoid any possibility of taking a fractional exponential power of a tiny
negative number due to truncation errors, which was leading to NaNs in some
cases while developing and debugging the other changes in this commit.  This
does not appear to change any answers in the existing test cases, perhaps
because the mixed layer restratification "tail" is not being activated by
setting TAIL_DH to be larger than 0.

  There was also the addition of code to both mixedlayer_restrat_Bodner and
mixedlayer_restrat_OM4 to determine the mixed layer thickness from its vertical
extent when in non-Boussinesq mode.  All answers are bitwise identical in
Boussinesq mode (at least when TAIL_DH=0.), but some solutions will change in
non-Boussinesq mode with this commit.