+*Non-Boussinesq wave_speeds wave structure calcs
  Modified the wave structure calculations in wave_speeds to avoid using the
Boussinesq reference density when in non-Boussinesq mode.  This includes the use
of thickness_to_dz, changes to the units of three arguments to wave_speeds and
their counterparts in int_tide_CS, and a number of duplicated calculations of
vertical extents mirroring the calculations of thicknesses.  Some diagnostic
conversion factors were modified accordingly in MOM_internal_tides.  There are 4
new internal variables and changes to the units of 3 other internal variables in
wave_speeds.  Several rescaling factors were cancelled out or replaced.  All
answers are bitwise identical in Boussinesq mode, but solutions with
INTERNAL_TIDES=True change in fully non-Boussinesq mode with this commit.